+++ src/gui/update.rs
@@ -314,24 +314,8 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
             );
             if let Some(path) = wow_path_resolution(chosen_path) {
                 log::debug!("Message::UpdateWowDirectory(Resolution({:?}))", &path);
-
-                if let Some(flavor) = flavor {
-                    // If a flavor is supplied we only update path for that specific flavor.
-                    let flavor_path = ajour.config.get_flavor_directory_for_flavor(&flavor, &path);
-                    if flavor_path.exists() {
-                        ajour.config.wow.directories.insert(flavor, flavor_path);
-                    }
-                } else {
-                    // If no flavor is supplied it will find as many flavors as possible in the path.
-                    let flavors = &Flavor::ALL[..];
-                    for flavor in flavors {
-                        let flavor_path =
-                            ajour.config.get_flavor_directory_for_flavor(flavor, &path);
-                        if flavor_path.exists() {
-                            ajour.config.wow.directories.insert(*flavor, flavor_path);
-                        }
-                    }
-                }
+                // Add directories
+                ajour.config.add_wow_directories(path, flavor);
 
                 // Clear addons.
                 ajour.addons = HashMap::new();
@@ -2066,15 +2050,15 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                 if let Some(wtf_folder) = ajour.config.get_wtf_directory_for_flavor(&flavor) {
                     return Ok(Command::perform(
                         list_accounts(flavor, wtf_folder),
-                        Message::ListWeakAurasAccounts,
+                        Message::ListWtfAccounts,
                     ));
                 }
             }
         }
-        Message::ListWeakAurasAccounts((flavor, result)) => match result {
+        Message::ListWtfAccounts((flavor, result)) => match result {
             Ok(accounts) => {
                 log::debug!(
-                    "Message::ListWeakAurasAccounts({}, num_accounts: {})",
+                    "Message::ListWtfAccounts({}, num_accounts: {})",
                     flavor,
                     accounts.len(),
                 );
@@ -2090,6 +2074,16 @@ pub fn handle_message(ajour: &mut Ajour, message: Message) -> Result<Command<Mes
                     [a] => Some(a.clone()),
                     _ => None,
                 };
+
+                // If there is only a single account, we save it to the config.
+                if let Some(single_account) = get_single_account() {
+                    ajour
+                        .config
+                        .weak_auras_account
+                        .insert(flavor, single_account);
+                    let _ = ajour.config.save();
+                }
+
                 if let Some(account) = account_from_config.or_else(get_single_account) {
                     if let Some(wtf_path) = ajour.config.get_wtf_directory_for_flavor(&flavor) {
                         state.chosen_account = Some(account.clone());